Bug Fix: Bad Creation Semantics on Named Semaphore
When we were attempting to create a semaphore that already existed, the
sem_open() call was failing, even when the O_EXCL flag was not set.

In this commit, I fix this bad semantic, so that in such a situation we
open the target named semaphore.

Note that I additionally have fixed the fault injection regression test.