Bug Fix: Sync with Servers
When the system is initializing, spawners progress in phases. At each
phase, they spawn the set of services that run in runtime level that is
being currently initialized, and then spawners sync.

Although we were syncing spawners correctly, we were messing up with the
initialization within the same runtime level. More precisely, we were
launching multiple servers at once, and then waiting their acknowledge.
Although this is semantically correct, if the underlying runtime system
features a cooperative thread implementation (and it does) we may face a
priority inversion problem.

In this commit, I fix this potential bug by forcing spawners to wait for
the acknowledge of a recently spawned server before spawning another
server in the same runtime level.