Bug Fix: Wait for Master Spawner on Shutdown
Spawners should synchronize with each other when shutting down. If not,
we are likely to abort the shutdown process of a service.

In this commit I introduce a workaround for this problem, by forcing all
spawners that do not implement the shutdown function to wait in an
infinite loop. This way, only those spawners that do implement a
shutdown routine will call exit. In the current implementation that
means spawner zero.