Feature: Polymorphic Messages
Feature: Polymorphic Messages

In this commit, I introduce a polymorphic structure for unnamed
mailbox messages. Now, there is a generic header structure that all mailbox
messages shall include. In this way, we split the message header independent of the implementation of each server.

Furthermore, I introduce some enhancements in the shutdown routine, so that it
uses a table for sending shutdown messages, instead of doing it verbosely.