Feature: Portal Sync Cache
In this commit, I introduce a caching mechanism for sync connectors that
are used by unnamed portals. Overall, instead of opening a sync
connector whenever we want to allow a remote to write into a portal, we
open it just once and keep its sync id. Later on, if we want to allow a
write operation from the same remote, we simply re-use the previously
opened sync connector.