Hotfix: check opening flags on shm services (#124)
* Weak client cache shm

* Bug fix: shm-server was not creating a shm with the correct oflags

In this commit, I correct the function call to create a shm on
the server, where previously the mode and writable arguments
were exchanged. The tests also had to be modified because they
were not passing the flags correctly, in short, they were opening
flags in the mode slot.

* Bug fix: checking on shm-server side

In this commit, I correct the use of the union of ret of shm_message
in the function unlink and I leave more explicit the operation reali-
zed in the return of the same one in the server.