Bug Fix: Use DMA Channels for Data Transfers
In this commit, I fix the use of DMA channels for data transfers in the
NoC. Before we were badly configuring DMA channels for mailboxes, and
for portals we were not setting up DMA channels at all. Furthermore, I
also removed the DMA channel configuration for sync connectors, because
it is kind useless.