Feature: Platform Independent Building
In this commit, I enhance the building process for the multikernel, so that it
is easier to hack the source tree to add the port for new targets. In summary,
I have introduced new compilation symbols that enables us to fine-select system
libraries to link agains when building a binary. Furthermore, I have also fixed
the source code, so that no platform dependent headers and symbols are placed
in platform independent code.